refactored dashboard UI and removed old task UI classes
